Sidecar - move SessionConfig, SidecarAction, and ApporQueue structs
Moving from interface.rs to separate files
