+++ sidecar/src/service/sidecar_server.rs
@@ -2,13 +2,13 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::config::get_product_endpoint;
-use crate::interface::{AppOrQueue, SessionConfig, SidecarAction, SidecarStats, TraceFlusher};
+use crate::interface::{SidecarStats, TraceFlusher};
 use crate::log;
 use crate::log::{MULTI_LOG_FILTER, MULTI_LOG_WRITER};
 use crate::service::{
-    InstanceId, QueueId, RequestIdentification, RequestIdentifier, RuntimeInfo, RuntimeMetadata,
-    SerializedTracerHeaderTags, SessionInfo, SidecarInterface, SidecarInterfaceRequest,
-    SidecarInterfaceResponse,
+    AppOrQueue, InstanceId, QueueId, RequestIdentification, RequestIdentifier, RuntimeInfo,
+    RuntimeMetadata, SerializedTracerHeaderTags, SessionConfig, SessionInfo, SidecarAction,
+    SidecarInterface, SidecarInterfaceRequest, SidecarInterfaceResponse,
 };
 use datadog_ipc::platform::{AsyncChannel, ShmHandle};
 use datadog_ipc::tarpc;
